# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Escolar",
        "description": "API para sistema de gerenciamento escolar multi-tenant. Permite cadastro de usuários, alunos, turmas, diários, eventos, entre outros. O contexto do tenant é definido pelo header 'x-tenant-id'.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "alexseyf66@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Servidor Local"
        },
        {
            "url": "https://elo-api-alexseyfs-projects.vercel.app",
            "description": "Servidor de Produção"
        }
    ],
...
    "paths": {
        "/login": {
            "post": {
                "tags": ["Autenticação"],
                "summary": "Login de usuário (Tenant)",
                "description": "Autentica um usuário de uma escola específica.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "senha": { "type": "string" }
                                },
                                "required": ["email", "senha"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Login realizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } } },
                    "400": { "description": "Dados inválidos" },
                    "401": { "description": "Credenciais inválidas" }
                }
            }
        },
...
        "/api/v1/platform/login": {
            "post": {
                "tags": ["Plataforma"],
                "summary": "Login de Administrador da Plataforma",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "senha": { "type": "string" }
                                },
                                "required": ["email", "senha"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Login realizado" },
                    "401": { "description": "Não autorizado" }
                }
            }
        },
...
        "/diarios": {
            "post": {
                "tags": ["Diarios"],
                "summary": "Criar entrada no diário",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": { "application/json": { "schema": { "type": "object" } } }
                },
                "responses": { "201": { "description": "Diário criado" } }
            }
        },
        "/diarios/aluno/{alunoId}": {
            "get": {
                "tags": ["Diarios"],
                "summary": "Diários de um aluno",
                "parameters": [
                    { "name": "alunoId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de diários" } }
            }
        },
        "/diarios/{id}": {
            "get": {
                "tags": ["Diarios"],
                "summary": "Obter diário por ID",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Dados do diário" } }
            },
            "put": {
                "tags": ["Diarios"],
                "summary": "Atualizar diário",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "requestBody": { "content": { "application/json": { "schema": { "type": "object" } } } },
                "responses": { "200": { "description": "Atualizado" } }
            },
            "patch": {
                "tags": ["Diarios"],
                "summary": "Atualizar diário (Parcial)",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "requestBody": { "content": { "application/json": { "schema": { "type": "object" } } } },
                "responses": { "200": { "description": "Atualizado" } }
            }
        },
        "/eventos": {
            "get": {
                "tags": ["Eventos"],
                "summary": "Listar eventos",
                "parameters": [
                    { "name": "data", "in": "query", "schema": { "type": "string", "format": "date" } },
                    { "name": "tipoEvento", "in": "query", "schema": { "type": "string" } },
                    { "name": "turmaId", "in": "query", "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de eventos" } }
            },
            "post": {
                "tags": ["Eventos"],
                "summary": "Criar evento",
                "parameters": [{ "$ref": "#/components/parameters/TenantId" }],
                "security": [{ "bearerAuth": [] }],
                "requestBody": { "content": { "application/json": { "schema": { "type": "object" } } } },
                "responses": { "201": { "description": "Evento criado" } }
            }
        },
        "/eventos/turma/{turmaId}": {
            "get": {
                "tags": ["Eventos"],
                "summary": "Listar eventos da turma",
                "parameters": [
                    { "name": "turmaId", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Lista de eventos" } }
            }
        },
        "/eventos/{id}": {
            "get": {
                "tags": ["Eventos"],
                "summary": "Evento por ID",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Dados do evento" } }
            },
            "put": {
                "tags": ["Eventos"],
                "summary": "Atualizar evento",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "requestBody": { "content": { "application/json": { "schema": { "type": "object" } } } },
                "responses": { "200": { "description": "Atualizado" } }
            },
            "patch": {
                "tags": ["Eventos"],
                "summary": "Atualizar evento (Parcial)",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "requestBody": { "content": { "application/json": { "schema": { "type": "object" } } } },
                "responses": { "200": { "description": "Atualizado" } }
            },
            "delete": {
                "tags": ["Eventos"],
                "summary": "Deletar evento",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
                    { "$ref": "#/components/parameters/TenantId" }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": { "200": { "description": "Evento deletado" } }
            }
        },
...
    },
    "components": {
        "parameters": {
            "TenantId": {
                "name": "x-tenant-id",
                "in": "header",
                "description": "Slug ou ID da escola (Tenant)",
                "required": true,
                "schema": { "type": "string" }
            }
        },
        "schemas": {
...
            "LoginResponse": {
                "type": "object",
                "properties": {
                    "token": { "type": "string" },
                    "id": { "type": "integer" },
                    "nome": { "type": "string" },
                    "email": { "type": "string" },
                    "roles": { "type": "array", "items": { "type": "string" } },
                    "schoolSlug": { "type": "string" }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}